Changed track number back to decimal so that we can handle partials.

+++ TagBot.Service/Sqlite.cs
@@ -122,7 +122,7 @@ namespace Tagbot.Service
                         Track temp = new Track
                         {
                             TrackName = (string)reader["title"],
-                            TrackNumber = (int)(Int64)reader["track_number"],
+                            TrackNumber = (decimal)reader["track_number"],
                         };
 
                         tracks.Add(temp);
