+++ TagBot.Service/models/Track.cs
@@ -6,7 +6,7 @@ namespace Tagbot.Service.models
 {
     public class Track
     {
-        public int TrackNumber { get; set; }
+        public decimal TrackNumber { get; set; }
         public string TrackName { get; set; }
     }
 }
